requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.0.0
pyarrow>=14.0.1
facebook-business==16.0.0
tableauhyperapi==0.0.22106
